fix: resolve 100+ SonarCloud issues across entire codebase
BLOCKERs fixed (4):
- TestDataGenerator: use try-with-resources for DB connections (java:S2095)
- TestDataGenerator: remove hardcoded password, use env vars (java:S6437)
- TestDataGenerator: reuse Random instance (java:S2119)
- JwtService: fix int overflow in expiration calculation (java:S2184)

Security fixes:
- ElasticsearchConfig: use JVM default SSLContext instead of trust-all (java:S4834)
- SecurityConfig: fix CORS wildcard+credentials combo violation (java:S5122)
- Remove @CrossOrigin(origins="*") from all controllers (CORS now global)

Code quality (80+ issues):
- Replace System.out/err with SLF4J Logger across all files (java:S106, ~53 issues)
- Replace @Autowired field injection with constructor injection (java:S6813, ~31 issues)
- Fix empty catch blocks with proper logging (java:S2139)
- Remove duplicate @Service annotation (java:S1118)
- Fix N+1 query: use findAllById+saveAll in deleteDishes
- Fix Dish entity ID type: Integer→Long for consistency with repository
- Fix Redis cache type-safety with instanceof check
- Upgrade JaCoCo 0.8.12→0.8.13 for JDK compatibility
- Remove unused imports and dead code
- Fix race condition in RecommendationService with volatile
- Add null checks in DishController and LlmService

+++ src/main/java/com/example/demo/model/Dish.java
@@ -1,19 +1,15 @@
 package com.example.demo.model;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import java.math.BigDecimal;
-
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "menu")
 
-public class Dish{
+public class Dish {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false,unique = true)
     private String name;
@@ -48,8 +44,8 @@ public class Dish{
 
 
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
